tests/wpt/tests/css/css-masonry/tentative/parsing/masonry-track-invalid.html

<!DOCTYPE html>
<title>CSS Masonry: masonry-track-* parsing</title>
<link rel="author" title="Sam Davis Omekara Jr." href="mailto:samomekarajr@microsoft.com">
<link rel="help" href="https://tabatkins.github.io/specs/css-masonry/">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
<body>
<script>
  test_invalid_value("masonry-track-start", "0");
  test_invalid_value("masonry-track-end", "span");
  test_invalid_value("masonry-track-end", "sPaN");
  test_invalid_value("masonry-track-end", '"1st"');
  test_invalid_value("masonry-track-end", "1st");
  test_invalid_value("masonry-track-start", "auto 1");
  test_invalid_value("masonry-track-start", "1 auto");
  test_invalid_value("masonry-track-end", "span 1 / span 2");
</script>
